+++ lib/form/fields/drop_down_field_with_modal_v2.dart
@@ -147,7 +147,7 @@ Future<T?> showSelectOptionModal<T>({
   String searchQuery = '';
   List<DropDownItem<T>> filteredOptions = options;
 
-  final isWeb = kIsWeb; // Verificar si es web
+  const isWeb = kIsWeb; // Verificar si es web
 
   if (isWeb) {
     return await showDialog<T>(