+++ lib/geo/domain/entities/geo_entity.dart
@@ -24,7 +24,5 @@ class GeoEntity {
   final String countryCode;
 
   @override
-  String toString() {
-    return name;
-  }
+  String toString() => name;
 }