+++ lib/user/models/user_phone_number.dart
@@ -1,11 +1,10 @@
 import 'package:equatable/equatable.dart';
-import 'package:hive/hive.dart';
+
 import 'package:json_annotation/json_annotation.dart';
 
 part 'user_phone_number.g.dart';
 
 @JsonSerializable()
-@HiveType(typeId: 7)
 class UserPhoneNumber extends Equatable {
   const UserPhoneNumber({
     required this.number,
@@ -18,11 +17,9 @@ class UserPhoneNumber extends Equatable {
   Map<String, dynamic> toJson() => _$UserPhoneNumberToJson(this);
 
   @JsonKey(name: 'number')
-  @HiveField(0)
   final String number;
 
   @JsonKey(name: 'cca2Country')
-  @HiveField(1)
   final String cca2Country;
 
   @override